refactor(ui): plugin add command trigger operation
