+++ packages/sheets-plugin-find/src/Controller/FindController.ts
@@ -9,7 +9,6 @@ import { FindModalController } from './FindModalController';
 import { FindMenuItemFactory } from './menu';
 
 export class FindController extends Disposable implements IDisposable {
-    private _findList: IToolbarItemProps;
 
     constructor(
         @Inject(FindModalController) private _findModalController: FindModalController,
@@ -20,31 +19,15 @@ export class FindController extends Disposable implements IDisposable {
         @ICommandService private readonly _commandService: ICommandService
     ) {
         super();
-        this._findList = {
-            name: FIND_PLUGIN_NAME,
-            toolbarType: 1,
-            tooltip: 'find.findLabel',
-            label: {
-                name: 'SearchIcon',
-            },
-            show: true,
-            onClick: () => {
-                this._findModalController.showModal(true);
-            },
-        };
         this._componentManager.register('SearchIcon', Icon.SearchIcon);
         const toolbar = this._uiController.getToolbarController();
-        // toolbar.addToolbarConfig(this._findList);
+
         this._initializeContextMenu();
 
         // TODO@Dushusir maybe trigger once in ui-plugin-sheets?
         toolbar.setToolbar();
 
         [ShowModalCommand, HideModalCommand].forEach((command) => this.disposeWithMe(this._commandService.registerCommand(command)));
-    }
-
-    getFindList() {
-        return this._findList;
     }
 
     private _initializeContextMenu() {
