+++ packages/sheets-plugin-import-xlsx/src/Controller/menu.ts
@@ -1,10 +1,11 @@
 import { IMenuButtonItem, MenuItemType, MenuPosition } from '@univerjs/base-ui';
 import { IAccessor } from '@wendellhu/redi';
 
+import { UploadCommand } from '../commands/upload.command';
+
 export function ImportMenuItemFactory(accessor: IAccessor): IMenuButtonItem {
     return {
-        // TODO@Dushusir import command trigger open file select
-        id: 'ImportCommand.id',
+        id: UploadCommand.id,
         type: MenuItemType.BUTTON,
         title: 'importXlsx.import',
         tooltip: 'importXlsx.tooltip',
